LLM: Make angle convention consistently [altitude, azimuth]

+++ src/test/java/rs117/hd/tests/TimeOfDaySnapshotTest.java
@@ -13,12 +13,12 @@ public class TimeOfDaySnapshotTest {
 		// update() is the per-frame entry point: it pins the instant every getter
 		// derives from and drops the previous frame's astronomy snapshot.
 		tod.update();
-		double[] first = tod.getSunAngles();
-		double[] second = tod.getSunAngles();
+		float[] first = tod.getSunAngles();
+		float[] second = tod.getSunAngles();
 		assertSame("within one frame, getSunAngles must return the cached array", first, second);
 
 		tod.update();
-		double[] third = tod.getSunAngles();
+		float[] third = tod.getSunAngles();
 		assertNotSame("update must invalidate the cache", first, third);
 	}
 }
